I was stupid, figured out where the kernel starts and ends

+++ include/system.h
@@ -36,8 +36,8 @@ typedef unsigned short u16;
 typedef signed short s16;
 typedef unsigned int u32;
 typedef signed int s32;
-typedef unsigned long u64;
-typedef signed long s64;
+typedef unsigned long long u64;
+typedef signed long long s64;
 
 typedef enum
 {
